+++ app/controllers/departments_controller.rb
@@ -69,6 +69,7 @@ class DepartmentsController < ApplicationController
     end
 
     def import
+        
     end
 
     private