+++ app/models/action.rb
@@ -1,0 +1,4 @@
+class Action < ApplicationRecord
+  belongs_to :perfil
+  accepts_nested_attributes_for :perfil
+end