+++ app/models/perfil.rb
@@ -1,4 +1,6 @@
 class Perfil < ApplicationRecord
-    validates :name, presence: true,
+    has_many :actions
+    accepts_nested_attributes_for :actions
+    validates :name, presence: true, uniqueness: true,
                      length: { minimum: 5}
 end