+++ db/migrate/20170914230919_drop_periods.rb
@@ -1,5 +1,0 @@
-class DropPeriods < ActiveRecord::Migration[5.0]
-  def change
-    drop_table :periods
-  end
-end